+++ utils/get-win32-envvar-from-wsl.js
@@ -1,7 +1,7 @@
 'use strict';
 
-module.exports = async (varname, {debug} = {}) => {
+module.exports = varname => {
   const args = ['-Command', `[Environment]::GetEnvironmentVariable('${varname}')`];
-  const {stdout} = await require('./run-command')('powershell.exe', args, {debug});
+  const {stdout} = require('./spawn-sync-stringer')('powershell.exe', args, {encoding: 'utf-8'});
   return stdout.trim();
 };